# src/main/java/com/idemia/http/dto/PersonDto.java
package com.idemia.http.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@NoArgsConstructor
@Data
@AllArgsConstructor
public class PersonDto {
    private PersonType type;
}